It was an overflow error!

+++ Project Euler/project7.h
@@ -5,8 +5,8 @@ class project7
 public:
 	project7(void);
 	~project7(void);
-	static std::vector<bool> sieve_of_eratosthenes(unsigned);
-	static unsigned find_cap(unsigned);
+	static std::vector<bool> sieve_of_eratosthenes(float);
+	static float find_cap(unsigned);
 	static unsigned find_nth_prime(unsigned);
 	static void runProject();
 };
